Replace ec2_ami_name with ec2_ami['Name']
Instead of just a name, we now specify a mapping with Name and
potentially other search criteria for AMIs.

The comment about ec2_ami_name is misleading—it's *always* needed, not
just when ec2_ami is defined; but if you specify AMIs by id explicitly,
we assume that you've also set the right user in the inventory. It isn't
the best assumption to make, but we can improve it later.