Comment out us-east-2 for now
It's still new, and there's no official Debian image for it yet (before
we restricted the search by owner, I think it was finding the image that
some random person had copied over from us-east-1!).

We'll surely have to add support for it in the very near future, but we
can figure out how later; by then, the "official" AMI may be present
anyway.