Use cluster_name everywhere, forget cluster_tags['Name']
I went through every use of both in the code, and although it would be
possible to change cluster_name to cluster_tags['Name'] in those places
where it's used to name an AWS resource rather than a file/directory, it
also seems pointless to expend the effort. We gain nothing from being
able to set these two values separately.

Instead, let's just (a) choose cluster_name so that it's suitable for
both AWS and filesystem use, and (b) use it everywhere and never set or
use cluster_tags['Name'] at all.

Fixes #101