Add a sys/firstboot role to handle early boot configuration
This is where we can configure the default user for the new AMI, but
right now it's just a bunch of notes. More research is needed here.

For now, we just remove /var/log/tpa-firstboot.log, so that instances
launched with the AMIs we build here don't have misleading output from
the initial run.

Towards #135