Switch to using a single long-lived S3 bucket
We have been creating one S3 bucket for each provisioned cluster, which
had numerous problems:

    (a) There was a race condition between IAM's propagation of the
    instance profile/role and the creation of the S3 bucket's policy
    referencing that role.

    (b) Setting the bucket lifecycle simply didn't work in certain
    regions (e.g., eu-central-1). No error, just failure.

    (c) We "optimised" the bucket for the first region mentioned in the
    cluster. Any changes to the region thereafter resulted in a slew of
    weird errors that took some unpredictable time to be resolved (e.g.,
    BucketAlreadyOwnedByYou in some regions only, or the dreaded "A
    conflicting conditional operation is currently in progress…").

    (d) There was also a race condition between DNS propagation for the
    bucket's name and awscli on the provisioned instances—sometimes the
    s3:GetObject request would receive 301 redirects with no location,
    or get other strange errors.

In summary, the S3 model really prefers to have long-lived buckets (but
the documentation doesn't actually say this in so many words).

So we create a 2ndquadrant-tpa bucket (optimised for no region) with a
one-day expiry policy (misc/s3-bucket.yml). You can specify a different
cluster_bucket in config.yml to override this.

We no longer try to create a bucket during provisioning, and we do not
use a bucket policy any more—instead, we add a statement to the instance
profile role's user policy that permits s3:GetObject on the cluster_name
prefix within the configured bucket. We no longer need to remember or
wait for instance_profile_role to be available.

Given the above, all the remains is simply to upload SSH host keys to
the correct prefix (ssh-hostkeys.yml), and fetch those using the awscli
(user-data/ssh-keys.yml).

Aside: this also means that we don't need to ask for ANY s3-related
permissions when someone's creating an IAM user for us to use.