Try to get awscli working as quickly as possible
On Debian and Ubuntu, there's the awscli package which is safe to use.
(It's much faster than trying to install python-pip, which has ~75MB of
dependencies.) On RHEL, the awscli package won't install because of an
unsatisfied dependency on python-docutils, but pip install is quick and
easy. In either case, we don't need curl and unzip.

Also, since we're mixing and matching installation methods, don't bother
to install awscli in common/pkg if it's been made available already.