Use any barman instance for barman-wal-restore
So far, if the immediate upstream had a defined backup, we could use it,
but in a cascading replication setup, that isn't sufficient any more. So
now we use any backup of the right instance (i.e., upstream_root).

Right now, the resulting decision is the same for all instances, and
does not try to be "clever" about which backup instance to use, if more
than one exists. We could do better: for example, we could prefer to use
backup instances in the same region. But that can come later.

Towards #134