Add support for parallel cascading replication setup
After much analysis, the only thing we really need to change is to
insert a dummy record into repl_nodes for any upstream that we're
supposed to replicate from that isn't already registered.

Also some bugfixes, most notably that an earlier commit had accidentally
inverted the logic for the test to decide whether to run repmgr standby
clone or not.

In passing, we also make the witness registration check be more cautious
and fix the same kind of problems there.

Many thanks to Gianni for working through the use case step-by-step with
me, and to Gülçin for help with analysis and testing.

Fixes #134