Get rid of trust from pg_hba.conf
We split postgres/createuser into postgres/createuser (applied to all
instances during configuration, generates and loads the password, but
does not create the user yet), and postgres/createuser/final (applied
to non-replicas, actually creates the user in postgres) and add a new
postgres/pgpass role to install ~postgres/.pgpass.

Note that we build up a list of users to create (on all instances) in
postgres_superusers (can be renamed to accommodate other users later, by
adding a role_attr_flags to the dict along with username), but we don't
use the list to generate .pgpass because we don't want the barman users'
passwords to be in ~postgres/.pgpass on other instances.