Add a final task to run pgbench
I've had this code kicking about for a while, and meant to improve it
somehow, but I've never figured out anything to do to make it better.
So here it is, unmodified.

Use «utils/deploy clustername --tags pgbench -e pgbench=run» to run it
(it's not run by default, though I suppose it could be, because it runs
only for 3+3 minutes; but this should probably be run along with various
other post-deployment tests from a separate playbook altogether).