Use flock instead of run_once in roles/secret
Using run_once tasks inside conditionally-included roles may be skipped
unexpectedly (if the host chosen for run_once happens to not fulfil the
condition), as described in ansible/ansible#18821. This has led to a
number of problems with roles/secret.

Instead, we get rid of run_once and use flock(1) to make sure that only
one of the hosts on which the task executes (i.e., those for which the
condition is true) will end up executing the command. (We would like to
use flock -E0 to make the processes that don't obtain the lock exit with
no error, but -E isn't supported in all versions of flock, so we use the
subshell trick instead.)