Changes to pg_hba.conf and postgresql.conf to accomodate barman backup
Addition of wal_level and replication slot related changes to
postgresql.conf of both coordinators and datanodes.

As of now, we are setting "trust" in pg_hba.conf for the barman,
streaming_barman and replication connections. I tried md5 since
we now have .pgpass for the barman users, but we would need to
copy them over to coordinator nodes as well. So going with "trust"
for now