Introduce a new postgres-xl/backup role for barman backup
In postgres-xl, we can have multiple datanode backends running
on the same instance. We can also have gtm/coordinator running
on the same instance. We need to backup these multiple backends
from the same instance. The postgres-xl/backup role will go
through the inventory and generate a new group of instances for
backup. We will generate a pseudo host corresponding to every
datanode that appears on each instance. They will be part of a
new 'backup_nodes' group. A separate play can then run on this
'backup_nodes' set of hosts to generate barman backup scripts
as needed.

Backup of GTM will be added separately. This only works with
coordinators/datanodes for now.