Move the add_host activity into a role of its own.
Also introduce addition of the GTM host. For the GTM host,
we pass a coordinator ip as a var. We also default the
"backup" variable to "" to avoid the barman related steps which
do not make sense for GTM. Instead, we use a gtm_backup variable
to point to the instance where backup of GTM will be done