Introduce the backup/gtm role to take GTM backups
The GTM backup activity can be configured via cron jobs. The
default is to run it every 4 hours.

This installs a shell script which first invokes a "CREATE BARRIER"
command in the XL cluster. It then backs up the contents of the
GTM directory into an appropriately timestamped directory location