Use only Cluster/node to identify instances
Deduplicating based on a whole slew of tags (cluster_tags plus instance
tags plus a few other bits) has the severe problem that a minor change
to some tag value can result in duplicate instances. Instead, we can use
the Cluster/node tags to identify instances.

(Aside: it's possible that we'll have to rename Cluster to TPA_Cluster
or something like that to avoid conflicts with customer tags, but it's
fine for now.)