Restrict cluster_name to letters, numbers, underscores
We want to be able to use the name directly in, e.g., repmgr.conf. It
would be nice to keep it lowercase so that it didn't need quoting at
all, but we already have MixedCaseNames in existing clusters. So we
accept [_a-zA-Z0-9]+ to keep anything else from sneaking in.