Acquire and release archive-wal lock before starting backup
If we invoke barman backup before the first segment has been archived
completely (i.e., the file shows up in wals/ AND the xlogdb has been
updated), it will fail. We can check for the file showing up, but we
don't know when the running barman archive-wal will finish.

So we acquire and release the archive-wal lock after we see that the
file has appeared in wals/, and invoke barman backup afterwards.

Closes #149