Set cluster_tag:vars in group_vars/ instead of hosts file
This commit makes several noteworthy changes:

1. It's no longer the responsibility of aws/provision.yml to generate
   the static inventory. Instead, it includes common/inventory/write.yml
   and sets vars appropriately.

2. The static inventory file contains only hostnames, IP addresses, and
   node numbers now.

3. All variable settings are under group_vars/ now, split into separate
   files (ansible settings, cluster settings, etc.).

The exact borders of the split (e.g., who sets which variables) are
liable to change, but the basic idea is that we hand instance_vars off
to inventory/write.yml, which generates the hosts file and writes vars
into group_vars.

This is part of a series of changes that are intended to make the static
inventory be the sole initial source of topology information, without
making anything important persistent via AWS tags, so that it can be
cleanly overriden by dynamic topology discovery. It's also related to
putting all generated information for the cluster in the inventory.