Write initial topology information to host_vars
We no longer write upstream, backup, and role to AWS tags at all. We do
write Cluster/node (used for exact_count/count_tag) and Name (because we
want to see the names in the AWS console), but any other tags that are
set in instances[].tags[] are just passed through to AWS, and not used
anywhere.

This makes the static inventory the single source of information about
the cluster's (initial, though we don't know that yet) topology. There
is no more confusion about whether ec2.py's idea of upstream should be
authoritative or not, because there's no variable translation going on
any more.

We still accept Name/upstream/backup/role under tags[] for backwards
compatibility, but they are removed from tags and set directly under
instances[].