Set default values for try_subkey()
For example, the following would not work:

    try_subkey([region,item,'az'])|default(omit)

because try_subkey would return None, which is considered to be defined
for the purposes of |default() (but not for |ternary()). There doesn't
seem to be a simple way to change the default to be jinja2's idea of
"null", so we just pass null as a default explicitly.

We also use more sensible/appropriate default values where possible,
which simplifies sanity checking later.

(Unfortunately, using a default of "null" for backup/upstream gives me
an error about finding a list instead of a string when trying to coerce
to Unicode. I don't know why, but it doesn't matter much, because we do
not check those values anyway, and we omit them from host_vars using a
|ternary() check, so it works fine anyway.)