Allow vars to be passed through for volumes[]
If you set vars: on the entries in an instance's volumes, they will be
written to host_vars for the instance. For example, putting the
following into config.yml

    - volumes:
        - device_name: /dev/xvdb
          …
          vars:
            mountpoint: /var/lib/postgresql

will result in host_vars/03-volumes.yml being produced with

    - volumes:
        - device: /dev/xvdb
          vars:
            mountpoint: /var/lib/postgresql

Nothing acts on these variables yet. The idea is that deploy.yml could
just iterate over the attached volumes and do something (such as invoke
the sys/fs role) for each one.

Note that we probably don't want to specify the mountpoint directly like
this, because it means we will need to change the configuration in more
than one place if we want to switch from Debian to RHEL, say. That's a
separate concern, however. This commit merely provides a mechanism for
information about volumes to be conveyed from config.yml to deploy.yml.
What information we specify and how it's used will be decided later.