Use postgresql_query instead of psql
This makes it easier to write the query (fewer worries about escaping)
and provides easier access to the results, with fewer error conditions
to worry about than when running psql.

This commit also cleans up some dubious logic around testing whether a
replica or witness is already registered as some other type of node. We
were selecting 'replica' and testing if it was != 'replica' as a sanity
check, which isn't the brightest idea.

(There's still a psql invocation for the INSERT into repl_nodes that I
haven't converted yet. Also, the bdr roles invoke psql directly in many
places, which I haven't looked at.)