Invoke sys/fs based on first volumes[] entry
!!! WARNING !!! This is a gross temporary hack.

What we should do is iterate through volumes[], invoking sys/fs on each
listed device, and either passing through any mountpoint specified, or
translating a "mount_as" to a platform-specific mountpoint, such as
the value of postgres_home.

What we do here is just enough to avoid some conditional logic to allow
debian-primary to have postgres_home on /dev/md0 while other instances
still have it on /dev/xvdb. It won't work for multiple volumes. It does
not honour the mountpoint or mount_as specified in config.yml.

But it's just enough to prove that the mechanism works.