Put tpa_version in the inventory
OK, I thought of a better way after all. If the value is in the
inventory, we don't need to lookup anything relative to cluster_dir. It
does mean the value won't be updated until utils/provision is run again,
but we're heading in that direction anyway, since we're writing vars
from config.yml to the inventory.