Add dstat to default packages
We already had everything installed (sysstat, basically) to complete the
tests described in Brendan Gregg's "Linux Performance Analysis in 60,000
Milliseconds" article:

https://medium.com/netflix-techblog/linux-performance-analysis-in-60-000-milliseconds-accc10403c55

That's the basic stuff. We want more. dstat is one of the few things
mentioned in the USE checklist that we don't already have:

http://www.brendangregg.com/USEmethod/use-linux.html

(smartctl is also listed, but doesn't make much sense on AWS.)