Use json_query to simplify the volume mapping
We have to cater to the possibilities that the ec2_vol_facts task was
skipped (because attach_existing was not set anywhere), or that it ran
and found nothing, or found something in some region and not in another,
and so on. We use a jmespath query to flatten the list of result volumes
and operate on them with a with_items loop.