Move ec2 key registration into a conditional include
The normal case is that ec2_key_name is defaulted in validate.yml, and
we register the key and use it when provisioning instances.

Some people incomprehensibly choose to avoid the key registration and do
the same thing by hand in user-data. We used to support this bizarre use
case with «--skip-tags ec2_keys», but this is inconvenient and requires
that the operator always remember to invoke it that way.

We now allow ec2_key_name to be set to "" explicitly in config.yml, and
avoid registering the key in this case. Then ec2_instance_key is never
set, and the key is omitted (as before) in the later invocation of the
ec2 module.

Since we write ec2_instance_key to vars.json, the other consequence is
that we skip key registration (and use the prior) value on subsequent
runs of provision.yml. That's desirable towards #218.