Remove overriden variables from provisioning_vars
There are various sources of variable settings:

    1. Explicitly set in config.yml
    2. Explicitly set in the playbook or on the command line
    3. Defaults calculated for the cluster by validate.yml
    3. Set in vars.json during provisioning

The previous change made it possible to set a provisioning variable like
ec2_instance_key (ordinarily set by ec2-register-key.yml) in config.yml.
We would then use the config.yml value instead of the vars.json one.

The problem is that the value would still remain in vars.json, and if we
subsequently removed ec2_instance_key from config.yml and reran, the old
value from vars.json would be used again. That may or may not be a good
thing, depending on the variable and how it's used.

It's safer to remove any overriden values from vars.json. We may end up
not skipping some tasks that we could have skipped, but that's better
than skipping them and using a stale value.