Allow bypassing iam tasks during provisioning
Like the previous change that allowed us to avoid «--skip-tags ec2_keys»
we now allow «--skip-tags iam» to be omitted based on config.yml
settings. If you set:

    instance_profile_name: xxx

provision.yml will not try to create the instance profile and attach an
inline policy at all—it will just use the name of the supplied instance
profile when invoking ec2.

If instead you set:

    cluster_profile: ""

then we'll skip the profile creation and omit the instance profile name
altogether when invoking ec2 (which is not a sensible thing to do).

Also, if you don't do any of the above and just run provision, it will
create the instance profile and store its name in vars.json, and skip
those tasks the next time (towards #218).