Abuse inventory_ignore_extensions to avoid annoying warning
Ansible 2.4 includes an .ini inventory plugin and happily feeds ec2.ini
(which was previously skipped) to it, resulting in this annoying warning
on every Ansible invocation:

 [WARNING]:  * Failed to parse /home/ams/work/2ndq/TPA/clusters/test/tpa/inventory/ec2.ini with ini inventory plugin: /home/ams/work/2ndq/TPA/clusters/test/tpa/inventory/ec2.ini:3: Expected
key=value host variable assignment, got: eu-west-1

 [WARNING]: Unable to parse /home/ams/work/2ndq/TPA/clusters/test/tpa/inventory/ec2.ini as an inventory source

It's not practical to move ec2.ini somewhere else, so we just set
inventory_ignore_extensions to 'ec2.ini' (it's meant to be a list of
file extensions to ignore, but the code is written so that it can be
any filename suffix, thank &deity).

This means that any files named "foo-ec2.ini" in the inventory directory
will be skipped, but there's a simple solution: don't create any such
files, and be thankful for the bright pink warning going away.