Allow utils/rehydrate to limit deployment targets
We need to split deploy.yml into two plays: one that runs against all
instances in the cluster and sets facts for them (e.g., hostvars, and
other facts that later roles depend on), and one that actually makes
changes.

We can't just use «--limit xyz» when invoking utils/deploy because we
need to run the fact discovery steps on hosts outside the limited set so
that we can make sense of upstream/backup references.

We make it possible to limit the effects of the second play by setting
the deploy_hosts variable on the command-line, and making rehydrate set
it to the instance that is being rehydrated.