Move postgres/vars invocation to the all-hosts play
We need "backup_name" set for all instances even if we're limiting the
changes to a single host, which means we need to move postgres/vars up
to the first play.

But postgres/vars depends on ansible_os_family, which means we need to
move fact discovery to the first play too. So we split some bits of the
common role out into separate roles that can be applied cluster-wide.

This leaves the common role as a bit of a relic: maybe its constituent
tasks should move to sys/pkg, sys/hostname, etc. But that's not really
important, just cosmetic.