Use «hosts: all» in deploy.yml
The purpose of using the named cluster tag was to mitigate the worst
effects of running the right deploy.yml against the wrong cluster (or
vice versa :-). Since we're moving towards a generic version that will
work against any cluster, we don't need to worry so much about that.

I would still have liked to be more specific than 'all' by writing

    hosts: "{{ groups[cluster_tag] }}"

But that gives me an error I don't really understand:

ERROR! the field 'hosts' has an invalid value, which appears to include a variable that is undefined.The error was: 'groups' is undefined

Worth investigating someday, perhaps, but this is OK for now.