Change the termination prompt handling
If we're handling multiple hosts—which we now permit—the "pause" module
issues only a single prompt (it seems to have an implicit run_once), and
the registered result is defined only on the host that it ran against.

I would love to have one confirmation prompt per host, but since that is
not working, we do the next best thing by not falling over and dying on
every other host.

(Aside: I suspect we would be better off removing this "Are you really,
really sure?"-style prompt entirely, but we can reconsider that later.)