Make cluster path handling more consistent for utils
All of the utils/ scripts that take a cluster name will now accept it in
two forms:

    1. An absolute path to a cluster directory, which is used as-is with
       no transformations.

    2. A relative path, either "x" or "clusters/x", in which case the
       absolute path to our "clusters/" directory will be prepended, and
       the result treated as if it were given as (1) above.

The path obtained thus is then checked to ensure that it exists.

The change is fairly straightforward for the scripts other than
ansible-cluster-playbook, which is a little too clever in accepting
either a cluster name or the path to a script within the cluster dir.
(Nevertheless, its behaviour is something we depend on, so we cannot
just make it be less clever.)