Remove the ec2 inventory cache during provisioning
The problem I was trying to solve in utils/rehydrate by calling ec2.py
--refresh isn't a problem only with rehydration, but with any fast
utils/provision invocation.

As such, it's better to fix it as part of provisioning, and it's best to
just delete the cache files to force ec2.py to reload its cache on the
next run, rather than trying to refresh immediately.