Add misc/find-unused-subnets
If you're writing a cluster and need some subnets that aren't already
used, you can just run

    misc/find-unused-subnets 3

and use the /24 subnets it prints. If you want a smaller subnet, just
take a /24 and split it up yourself.

We also had a medium-sized blob of not-very-documented Python code to do
something like this, but not quite useful. Out it goes.