Remove unnecessary/old testing clusters
platforms/ goes away because (a) we don't actually have much need for
mixed-platform clusters, (b) it's very difficult to manage with one
named AMI and a bunch of ami-xxx id references, and (c) we can easily
modify test/tpa to allow a different AMI name to be specified on the
command line to spin up an RHEL cluster instead of Debian.

bdr2_rhel goes away for the same reason, and bdr because it's nearly the
same as bdr2, but older. xl goes away because it's untested and unused,
not because it's not something we need. We'll certainly revisit this in
future, especially once we have XL packages.

If anyone needs an XL cluster for testing in the meanwhile, it's easy
enough to dig this up again, but some extra work _will_ be needed to
adapt to the current TPA code.