Wait for instance reachability checks to pass before testing ssh
Most instances come up quickly, but sometimes it takes several minutes
(we don't know why), and in some cases the initial ssh-port-open test
seems to pass but the instance is not yet initialised/reachable, and
subsequent tasks fail.

We can be extra careful by repeatedly querying the ec2_instance_status
of each instance until system and instance status are both ok, and the
reachability tests in particular have passed. (Is that enough? We don't
know yet.)

Many thanks to Marc for help tracking this down and deciding how to
approach the problem.