Move hostvars and local under roles/platforms
We now have a "platforms" wrapper role that invokes platforms/aws on EC2
instances, and platforms/common everywhere. The former hostvars.yml are
absorbed therein, and the ec2-specific code in local/tasks/wait.yml is
now applied only to EC2 instances.

The short-lived local role goes away altogether now: the binary checks
are performed in platforms/tasks, and we're not enabling sudo/become in
this first play, so the tasks don't need to be exempted with become: no.
The last step here is the (cross-platform) SSH wait_for.

This simplifies deploy.yml: now we just apply three roles to every host
at the beginning: platforms, facts, and postgres/vars.

No more ../../wait/was/that/too/many/dots/hostvars.yml!