Remove other ec2_ references outside roles/platforms
These references in postgres-xl are supported for now through a legacy
XL hostvars translator, but they should be migrated to the inventory in
the future and not set as AWS tags at all. We can also use nicer names
now that the prefix is removed.

This means there are no remaining references to ec2_* outside
roles/platforms, which is the way we should strive to keep it.