Increase the default ansible connection timeout
The default of 10s is usually fine when you're using Ansible on a fast
network, but when speaking to remote AWS servers in particular, a much
longer timeout is regularly called for. 120s may be slightly excessive,
but it's come in handy a couple of times already.

This value can be overriden by setting ANSIBLE_TIMEOUT=nnn in the
environment or by adding -Tnnn to the ansible command line.