Fix cluster_network handling
If we're not using sys/openvpn (which will set cluster_network before we
try to create pg_hba.conf, whether we had it set before or not), we need
to set cluster_network explicitly.

So we add code to check that it's set, discover the earlier code was not
looking in cluster_vars at all, and shuffle things around a bit to make
it work properly. As a convenience, we also set cluster_network for
single-subnet clusters.