Set TPA_DIR/tpa_dir to TPA path everywhere
This will let us not have to use so many relative paths in playbooks and
roles. We export TPA_DIR in utils/ansible and use that to set tpa_dir in
validate.yml (for provisioning) and platforms/vars (for deployment). As
an extra caution, both sites can fall back to using their own location
to derive the parent (well, ancestor) path too.