Make ansible-cluster-playbook take an optional playbook argument
Now we can do this:

    utils/ansible-cluster-playbook tpa misc/find-volumes.yml

Note that the argument must resolve to a filename, i.e., either be an
absolute path, or relative to the cwd. You can't specify the name of a
playbook that's within clusters/tpa, for example. If you want that, the
earlier syntax of tpa/foo.yml still works.

We also remove the default script=deploy.yml. It's hard to imagine that
someone who runs «ansible-cluster-playbook xyz» would want to run deploy
by default.