Add a 'TPAAttachmentLabel' tag alongside Name for EBS volumes
Relying on the Name of a volume to follow a particular pattern turns out
to be unreliable, because the Name may be silently changed by background
processes in certain customer environments.

So in addition to the Name, we set another tag that we can use to search
for volumes. Of course, none of the existing volumes will have this tag,
but we can run misc/tag-volumes.yml to create them. (Even if this step
is missed, we now fail in provision if suitably-tagged volumes weren't
found, so the consequences are limited.)

We also make the sanity check for volume tags during rehydration a
little more specific/verbose about what the problem is.