Don't hardcode 'become: postgres'
One the one hand, we want to use 'postgres', but we do have the
postgres_user setting, and if we change it, all of these become
directives would break. On the basis that automated deployments may have
legitimate reasons to use different usernames (e.g., colocated XL nodes)
we use "{{ postgres_user }}" consistently.