Use become_user/become instead of sudo_user/sudo
The latter form is deprecated, and using become_user/sudo (which is more
readable than the hideous «become: yes») is no longer allowed in 2.4. So
we might as well be consistent.

(I considered using sudo_user/sudo everywhere, but we use dzdo as the
become_method in at least one environment, so it would be misleading.)