Revise cluster_discovery to return basic facts
The caller of the module provides inputs in two ways: it invokes the
module as a particular user (postgres_user) and provides a conninfo
string to connect to Postgres as a superuser.

I had some thought of passing in whatever we knew to the module, but it
turned out that wasn't necessary. Given the two inputs above, everything
else can be discovered using that as a starting point.