Tweak instance status module and conditions
I assumed that DescribeInstanceStatus would always return something for
an existing instance, but we may get InstanceStatuses: [] instead. The
earlier code would die trying to retrieve InstanceStatuses[0], and the
loop would never terminate because i|succeeded would never be true.

So we adjust the condition: terminate the loop if the request failed or
passed, but try again if the request didn't fail, but we either got no
results, or positive results indicating that the tests haven't passed.
This needs some small changes to the module code, which progressively
replaces status at each stage and adds 'request_status': 'ok' when we
have a response to test.