Set active= based on whether /var/lib/barman/name exists
This allows us to gracefully handle the case where /var/lib/barman/name
exists but /etc/barman.d/name.conf does not (because we reattached the
barman_home volume formerly attached to a working Barman instance).

Another way to do this would be to parse barman diagnose output and use
that to determine what remedial actions are required for the backups of
a particular instance. That may be required with more complex backup
setups, but for now, this simple approach suffices.