Add pg_isready after starting/restarting postgres
This is preferable to assuming that Postgres has been started and then
failing somewhere far down the line when we discover that it has not.

A server that enters recovery on startup may take considerably longer
than a minute to start accepting connections, so this task will fail.
But that's still preferable to carrying blithely on in ignorance, and
it's safe to re-run deploy anyway.