Install a custom repmgr.service file
Behaves the same way on Debian and RedHat, runs repmgr in the foreground
(more systemd-friendly), and avoids the problem of /var/run disappearing
on boot (we created it in repmgr/pkg, but didn't add it to tmpfiles.d).

I would have liked to call this 'repmgrd.service', but a systemd bug on
(at least) Debian 8 makes it impossible to cleanly override the service
unit intuited from /etc/init.d/repmgrd, and using a different name is
the only easy workaround available (based on #systemd discussions).