Replace postgres-monitor with a Python script
pg_isready is documented to return 1 if it says 'rejecting connections',
but I saw it apparently returning 2 for that case and therefore exiting
early while the server was in recovery. Instead of being subject to its
vagaries, we write the same logic using psycopg2, which also allows us
to be more selective about which errors we consider acceptable.

(I don't understand how pg_isready can possibly return 2 when it gets
PQPING_REJECT and says 'rejecting connections'. It's possible that the
cause of the error I observed lies elsewhere… but the fact remains that
the shell script didn't work, and this one does.)