Apply postgres/config to all instances
On a primary instance, it continues to do what it used to before. On a
replica, it assumes that the basic configuration has been cloned from
the primary, and drops in replica override files that contain only any
replica-specific settings that differ from the upstream configuration.

We also have to include the replica configuration (using includes and a
bit of hackery, rather than include_role/tasks_from) after we run repmgr
standby clone (because postgres/config was already applied before the
data directory on the cloning instance existed, and we don't want to
wait for the next deployment to override the configuration).

XXX It may be possible to do something more elegant and less repetitive
about generating the configuration files, and especially about quoting
the values, but what we have works for now.