Use cluster_facts to improve repmgr/replica/final
Earlier, we would check if the instance had been registered with repmgr
as a replica, and if not, run repmgr standby clone and repmgr standby
register.

We can now make this procedure more robust by using cluster_facts. We
now ask a series of questions and take specific steps to correct the
problems we may discover while answering them. For example:

* Do we have a valid PGDATA?
  No: repmgr standby clone

* Is the instance registered?
  No: repmgr standby register

The new structure gives us a sound basis to build more sanity checks,
for example, checking for and fixing replication slot problems.

We also use postgres-monitor to wait for Postgres to start accepting
connections instead of the pg_isready earlier. Even this may fail, but
of course it's safe to rerun the whole thing. In particular, we don't
need to remove earlier-cloned PGDATA to make the right thing happen.