Drop unused slots after repmgr setup
When we clone a replica from an instance that is not the one it will be
streaming from eventually, repmgr3 leaves an inactive slot behind after
the clone has finished (the slot is required during cloning, of course).
We're quite conservative about it, but we can now remove that slot after
the replicas have been set up.

This is the first step towards #215