Make postgresql_query accept 'query' xor 'queries'
If we declare query as type='list' and pass in a string, it's split on
commas and passed in as a list, so this breaks horribly:

    - postgresql_query:
        query: >
            SELECT 41, 42, 43

There's no sane way to handle this, unfortunately, so we make the module
take a single query as a string (query=xxx, which is not split up), or
queries as a list (queries=[xxx, yyy, …]). The latter form also allows
the query text and arguments to be separated.

Now we can do any of the following:

    - postgresql_query:
        query: >
            SELECT xxx, yyy, zzz …

    - postgresql_query:
        queries:
            - SELECT 42 as a
            - SELECT 43 as b
            …

    - postgresql_query:
        queries:
            - SELECT 42 as a
            - text: SELECT 40+%s as b
              args:
                - 3

(Note that you have to use 'queries' even if you want to execute a
single query with text and args separate. A pity, but there's no easy
workaround that would allow query to be used instead in that case.)