Rename 'mount_as' to 'volume_for'
There are two ways to declare the intended use of a volume.

    (a) Specify a mountpoint directly

        volumes:
          - device_name: /dev/xvdb
            …
            vars:
              mountpoint: /foo/bar

    (b) Specify a label that says what the volume is for:

        volumes:
          - device_name: /dev/xvdb
            …
            vars:
              volume_for: foobar_data

Using 'mount_as' for the latter was not appropriate, because initdb will
not allow a mountpoint to be used as PGDATA directly, so it's not really
"_mount_ as", it's more like the "mountparent". So we call it volume_for
because it declares what the volume is for. That we derive a mountpoint
from the declared purpose is a detail.

(All of this is less important now that we no longer try to adapt to the
distribution's idea of where PGDATA should live. The original idea was
for «mount_as: postgres_data» to be translated to postgres_data_dir on
the fly, but now we might as well set «mountpoint: /opt/postgres/data»
everywhere. Still, I have a feeling the mechanism might come in handy
somewhere, so I'm leaving it in for now.)