Always execute openvpn from $PATH
Using a hardcoded absolute location doesn't work across Linux and MacOS
X, so we depend on $PATH instead. Of course, openvpn isn't likely to be
in the PATH on either system, so we add a couple of likely locations to
PATH first and check that we can find the binary somewhere.

Based on a report from Sasha in #229