Skip installing known_hosts if there's no hostkeys/ directory
For some reason that is not clear to me, and is not documented in the
proposed commit, GDS apparently does not have a hostkeys/ subdirectory.
Without having the contents of the host keys locally, there is no way to
generate known_hosts that is fast enough for a large number of hosts (as
seen in XL deployments, for example). So we just skip it altogether.

(I have not considered anything outside the scope of this immediate
request, but there is other local state used during deployment that
isn't so easily skipped, and I have no idea how GDS deals with it.)

Based on PR #229 from Sasha