Always rewrite host_vars/x/03-volumes.yml
Earlier, we would write the file only if one or more of the volumes for
the instance had 'vars' defined. This ran the risk of failing to clear
a file written earlier if the vars were later removed.

There's no advantage to omitting the file, so instead we just write it
always. Even if there are no vars defined (or volumes defined), we can
sort it out later in sys/volumes.